requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
strawberry-graphql[asgi]==0.235.1
httpx==0.27.2

# exact pydantic stack the app expects
pydantic==2.9.2
pydantic-settings==2.6.1
jsonschema==4.23.0

# Outlook .msg parsing (pin to avoid conflict)
extract-msg==0.42.0
olefile==0.46

# PDFs & images
PyPDF2==3.0.1
Pillow==10.4.0
pdf2image==1.17.0

# OCR: Tesseract (system binary) + PaddleOCR (Python)
pytesseract==0.3.13
paddlepaddle==2.6.1
paddleocr==3.3.0

# Storage / misc
boto3==1.35.21
python-multipart==0.0.9
python-magic==0.4.27
filetype==1.2.0

# auth
PyJWT>=2.8

# PII / logging / metrics
presidio-analyzer==2.2.357
presidio-anonymizer==2.2.357
structlog==24.4.0
prometheus-client==0.20.0

# Telemetry (optional)
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-httpx==0.46b0

# AV & cache / sessions
clamd==1.0.2
redis==5.0.8

# Local VLM/LLM via Ollama
ollama==0.6.0
